rename mermaid.js to Mermaid.js
